requirements.txt
chromadb
chainlit
fastapi
uvicorn
python-docx
pymupdf
torch
sentence-transformers
